Improve heuristics for finding the user frame that dispatched a Redux… (#10576)
* Improve heuristics for finding the user frame that dispatched a Redux action

The original logic was iterating backwards across the point stack frames looking for an applyMiddleware frame, and going backwards one more frame from there. This is problematic because I'm seeing variance in whether or not applyMiddleware actually shows up in stack frames. For some reason it shows up in Metabase recording stack traces, but not in our own E2E tests (and we're both using RTK).  That meant that the existing logic could keep right on walking backwards long past the frame it "should" have found.

What I've settled on is using the logic we had that tries to identify "is this function in a Redux middleware" based on source outlines (looking for the triply-nested function signature of a middleware definition), and stopping at the first frame that appears to be not a middleware and not applyMiddleware.  This seems to be producing pretty reasonable results.

* Update to a recent `breakpoints-01` recording

* Force re-running the J2C routine

* Add a util for verifying J2C behavior

* Add a test for Redux J2C behavior

* Ignore wrong lint error

* Update RDT-02 test to match current example recording