Update `experimental` from latest `main` (#10590)
* Jump-to-source falls back to parent user event if marker frame cannot be found (#10561)

### [Loom overview](https://www.loom.com/share/6d697c9136274fb8892325d39283fdac)

Looking at [go/r/bb82f5c3-af5c-4e20-9f1c-5efa4ee6bd29](https://app.replay.io/recording/appreplayio--bb82f5c3-af5c-4e20-9f1c-5efa4ee6bd29), I observed that the `getCypressMarkerFrame` lookup fails to find a "marker frame" because there are only two sources in the call stack– the Replay plugin itself and Cypress. That method expects to find user code in between for some reason, so this breaks the jump-to-source behavior.

In this case, the source is:
```js
cy.contains('Product added to cart!').should('be.visible');
```

That gets rendered in the UI as this:

![Screenshot 2024-06-13 at 10 46 16 PM](https://github.com/replayio/devtools/assets/29597/63920c23-82c8-45b4-9bb9-29443984b3b9)

Clicking on the failed _assert_ (or trying to "jump to source") would ideally jump to the `should('be.visible')` but it fails because of the "marker frame" issue I mentioned above. The idea occurred to me that in this case, jumping to the parent command (`contains('Product added to cart!')`) is probably better than nothing. (In this case, they're even on the same line, but that won't always be true.)

* Add data-private attributes (#10562)

* Remove [data-private] CSS style (#10563)

* Reset selected test event when test changes (#10565)

* Improve visual style for React/Reduxt not-loaded error messages (#10566)

* Log point panel should never show 1/0 (#10568)

* Fix video toggle for horizontal DEV layout (#10569)

* Support multi-line print statements (#10570)

Support multi-line print statements to improve readability and text editing experience.

* Stop mousedown event propagation in the Focuser (#10575)

* Adjusting contrast (#10572)

* Always use hoverTime for the screenshot when the user is hovering (#10578)

* Fix detection of async parent pauses outside of the focus window (#10573)

* Revert "Improve message shown when async stack can't be loaded (#10557)"

This reverts commit 585e560d741d2db812010bc35fe442af8b25a7df.

* Fix detection of async parent pauses outside of the focus window

* Remove LogRocket from Devtools (#10577)

* Improve heuristics for finding the user frame that dispatched a Redux… (#10576)

* Improve heuristics for finding the user frame that dispatched a Redux action

The original logic was iterating backwards across the point stack frames looking for an applyMiddleware frame, and going backwards one more frame from there. This is problematic because I'm seeing variance in whether or not applyMiddleware actually shows up in stack frames. For some reason it shows up in Metabase recording stack traces, but not in our own E2E tests (and we're both using RTK).  That meant that the existing logic could keep right on walking backwards long past the frame it "should" have found.

What I've settled on is using the logic we had that tries to identify "is this function in a Redux middleware" based on source outlines (looking for the triply-nested function signature of a middleware definition), and stopping at the first frame that appears to be not a middleware and not applyMiddleware.  This seems to be producing pretty reasonable results.

* Update to a recent `breakpoints-01` recording

* Force re-running the J2C routine

* Add a util for verifying J2C behavior

* Add a test for Redux J2C behavior

* Ignore wrong lint error

* Update RDT-02 test to match current example recording

* Log point panel saves pending edits when removing condition (#10579)

* Add option to notify user when processing is complete (#10582)

* Guard against undefined value (#10581)

* Source viewer font-size fixes (#10580)

Print statement panel dynamic sizing bug fixes:
- Fixed minor font-size and line-height inconsistencies for print statement panel
- Handle edge case with empty last line (aka Shift+Enter)

![Screenshot 2024-06-21 at 10 58 36 AM](https://github.com/replayio/devtools/assets/29597/1a8e9514-1c56-4998-84e0-7b71a1d022b7)

Also updated the `useSourceListCssVariables` hook to reevaluate CSS variables (width of line numbers and hit counts) when the font size preference changes.

* Better view element icon (#10583)

* Enable asserts (#10584)

* Handle lack of `point.frame` (#10586)

* Remove unused `jumpToClickEventFunctionLocation` (#10585)

* Improve video cursor display, especially for Cypress tests (#10587)

* Extract utils to check for user click/keyboard test steps

* Simplify UserActionEventRow J2C checks

* Use mouse event times for hover display for click steps

* Narrow time that we show a mouse click indicator

* Only show mouse clicks within the current focus window

* Revert focus window filtering for mouse cursors

* Update the Cypress steps test to check for hover cursor position (#10588)

* Update cypress/bankaccounts.spec example

* Add test for Cypress cursor positioning

* Use dashed/camel-cased data names

---------

Co-authored-by: Brian Vaughn <brian.david.vaughn@gmail.com>
Co-authored-by: Holger Benl <hbenl@evandor.de>
Co-authored-by: Jon Bell <github@lot23.com>
Co-authored-by: Domi <Domiii@users.noreply.github.com>
Co-authored-by: Mateusz Burzyński <mateuszburzynski@gmail.com>